requirements.txt
azure-ai-vision-imageanalysis==1.0.0
azure-core==1.32.0
certifi==2025.1.31
charset-normalizer==3.4.1
idna==3.10
isodate==0.7.2
numpy==1.24.3
psycopg2==2.9.10
pygame==2.6.1
python-dotenv==1.1.0
requests==2.32.3
six==1.17.0
typing_extensions==4.13.0
urllib3==2.3.0
